+++ crates/tools/tool-http-server/src/main.rs
@@ -1,7 +1,5 @@
 use crate::cli_utils::initialize;
 use crate::http_app::start_server;
-use common_cli::tool_exit_helpers::exit_error;
-use tracing::error;
 
 mod cli_utils;
 mod http_app;
@@ -9,15 +7,7 @@ mod models;
 
 #[tokio::main]
 async fn main() {
-    let args = initialize();
+    let config = initialize();
 
-    match args {
-        Ok(a) => {
-            start_server(a).await;
-        }
-        Err(e) => {
-            error!("{}", e);
-            exit_error()
-        }
-    }
+    start_server(config).await;
 }